refactor: move types to components package
