+++ internal/tui/components/hpicker_keys.go
@@ -1,4 +1,4 @@
-package hpicker
+package components
 
 import (
 	"github.com/charmbracelet/bubbles/key"
