+++ internal/tui/components/hpicker_styles.go
@@ -1,4 +1,4 @@
-package hpicker
+package components
 
 import "github.com/charmbracelet/lipgloss"
